[FIX] product: Move onchange for Decimal Accuracy and UOM to product
The commit
97430bba609882cd734b79f89f903d7894266145
set 2 warnings on Units of Measure and Decimal Accuracy.
However this introduced a bug, here are the steps to reproduce:
- Install decimal_precision
- In debug mode, go to Decimal Accuracy
- Create: traceback

Based on commit
916df82eba1a98cd5a6c780f2e7ca16356fa640c
we move one warning from decimal_precision to product, and modify the
translations accordingly.

X-original-commit: a6a83a4539a17ff9a3078a385b0e147000cd9727
