+++ addons/decimal_precision/i18n/decimal_precision.pot
@@ -110,17 +110,3 @@ msgstr ""
 msgid "Usage"
 msgstr ""
 
-#. module: decimal_precision
-#: code:addons/decimal_precision/models/decimal_precision.py:24
-#, python-format
-msgid "Warning!"
-msgstr ""
-
-#. module: decimal_precision
-#: code:addons/decimal_precision/models/decimal_precision.py:26
-#, python-format
-msgid "You are setting a Decimal Accuracy less precise than the UOM:\n"
-" %s \n"
-"This may cause inconsistencies in reservations.\n"
-"Please increase the rounding of this unit of measure and the global decimal precision."
-msgstr ""
