+++ addons/product/models/__init__.py
@@ -5,6 +5,7 @@
 
 from . import res_config_settings
 from . import decimal_precision
+from . import uom_uom
 
 # don't try to be a good boy and sort imports alphabetically.
 # `product.template` should be initialised before `product.product`
